GUI: - Give Options Tab its own class. - Play around some more with the paperdoll. Hitting simulate in the paperdoll tab now does a normal simulation, creating a html report.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16675 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ qt/simcpaperdoll.cc
@@ -274,7 +274,7 @@ PaperdollProfile::setProfession( int profession, int type )
 void
 PaperdollProfile::setSpec( int player_spec )
 {
-  assert( player_spec <= dbc::specialization_max_per_class() );
+  assert( player_spec <= (int) dbc::specialization_max_per_class() );
   m_spec = dbc::spec_by_idx( m_class, player_spec );
 
   for ( slot_e t = SLOT_INVALID; t < SLOT_MAX; ++t )
